event-system #17: AR-5 fix pass (2M+1L)
M-1 EventBus.PublishAuthoritative: structSize <=0 / > MaxEventSlotBytes
guards reordered to precede QueueCount++ slot reservation. If a guard
threw after QueueCount++ (both are registration errors but recoverable
in some Stage 1 hosts), the slot was reserved but PayloadBuffer/SlotMeta
never populated; subsequent Publish wrote to slotIndex+1, leaving slot N
with EventTypeOrdinal=0 (CLR default), which SerializeLedger classifies
as Tier A (GetTier(0)==0) and emits as a zero-byte record, corrupting
the canonical FM-017-001 digest.

M-2 EventLedger.Subscribe: hard cast (EventTypeDispatcher<T>) replaced
with 'as' + null-check, emitting ERR_EVT_ORDINAL_COLLISION (0x1707)
diagnostic on duplicate-ordinal RegisterExternalRow misuse. Symmetric
to AR-4 L fix in CosmeticChannel.Subscribe; the Tier A/B path was
missed and would still throw an opaque InvalidCastException with no
ordinal context.

L-1 CardIssuedEvent.FoulOrdinal: widened byte -> ushort. Doc claimed
it stored "IntraPhaseDrawIndex of the FoulCommittedEvent that triggered
this card", but intraPhaseDrawIndex is a ushort (range 0..65535) per
§2.4.1 -- a byte field could not represent draw indices > 254. Sentinel
0xFF -> 0xFFFF. No call sites exist yet (Stage 1 wiring not implemented),
so widening is non-breaking.

Files: EventBus.cs v1.5, EventLedger.cs v1.4, CardIssuedEvent.cs v1.2.

+++ src/event-system/CardIssuedEvent.cs
@@ -38,14 +38,16 @@ namespace TacticalDirector.EventSystem
         public readonly int Recipient;
         /// <summary>Card kind: 0=Yellow, 1=Red, 2=SecondYellow (domain ordinal).</summary>
         public readonly byte CardKind;
-        /// <summary>IntraPhaseDrawIndex of the FoulCommittedEvent that triggered this card; 0xFF if procedural (no associated foul).</summary>
-        public readonly byte FoulOrdinal;
+        /// <summary>IntraPhaseDrawIndex of the FoulCommittedEvent that triggered this card; 0xFFFF if procedural (no associated foul).
+        /// AR-5 L-1: widened from byte to ushort to match the ushort intraPhaseDrawIndex
+        /// header field (range 0..65535) — a byte sentinel could not represent draw indices &gt; 254.</summary>
+        public readonly ushort FoulOrdinal;
 
         /// <summary>
         /// Constructs a <see cref="CardIssuedEvent"/> with payload fields only.
         /// Header fields are left at zero — EventBus.Publish will overwrite them.
         /// </summary>
-        public CardIssuedEvent(int recipient, byte cardKind, byte foulOrdinal)
+        public CardIssuedEvent(int recipient, byte cardKind, ushort foulOrdinal)
         {
             eventTypeOrdinal    = 0;
             payloadVersion      = 0;
@@ -64,4 +66,10 @@ namespace TacticalDirector.EventSystem
 // | Version | Date       | Author | Notes                                                                   |
 // | 1.0     | 2026-05-30 | —      | Initial implementation.                                                 |
 // | 1.1     | 2026-05-30 | —      | AR-1 L-1: FoulOrdinal doc corrected (-1 invalid for byte; now 0xFF).   |
+// | 1.2     | 2026-06-02 | —      | AR-5 L-1: FoulOrdinal widened from byte to ushort. Doc claimed it      |
+// |         |            |        | stored "IntraPhaseDrawIndex of the FoulCommittedEvent that triggered   |
+// |         |            |        | this card", but intraPhaseDrawIndex is a ushort (range 0..65535) per   |
+// |         |            |        | §2.4.1 — a byte field could not represent draw indices > 254.         |
+// |         |            |        | Sentinel updated 0xFF → 0xFFFF. No call sites exist yet (Stage 0       |
+// |         |            |        | wiring not implemented), so widening is non-breaking.                 |
 #endregion
